Changes needed to make "all" build complete, including fixing the generated converter tool. Review URL: http://codereview.chromium.org/173045
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23721 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 76b3af422d58822e17523064aca0a3de116a3457

+++ o3d/core/cross/bitmap.cc
@@ -344,11 +344,11 @@ void Bitmap::DrawImage(const Bitmap& src_img,
   DCHECK(src_img.image_data());
   DCHECK(image_data());
 
-  if (dst_level < 0 || dst_level >= num_mipmaps()) {
+  if (dst_level < 0 || dst_level >= static_cast<int>(num_mipmaps())) {
     O3D_ERROR(service_locator()) << "Destination Mip out of range";
   }
 
-  if (src_level < 0 || src_level >= src_img.num_mipmaps()) {
+  if (src_level < 0 || src_level >= static_cast<int>(src_img.num_mipmaps())) {
     O3D_ERROR(service_locator()) << "Source Mip out of range";
   }
